+++ assets/CircularProgressPage.4fd09225.js
@@ -1,0 +1,13 @@
+var t=Object.defineProperty;var e=(n,s)=>t(n,"name",{value:s,configurable:!0});import{C as o}from"./CircularProgress.8fb9c828.js";import{C as a}from"./ComponentInfo.6570b779.js";import{S as i}from"./copyText.16911af4.js";import{d as r}from"./index.a47904f5.js";import"./Link.0ea0b328.js";import"./PaperCode.b42d056f.js";function c(){return r(i,{sx:{color:"grey.500"},spacing:2,direction:"row",get children(){return[r(o,{color:"secondary"}),r(o,{color:"success"}),r(o,{color:"inherit"})]}})}e(c,"CircularColor");c.code=`import CircularProgress from "@suid/material/CircularProgress";
+import Stack from "@suid/material/Stack";
+
+export default function CircularColor() {
+  return (
+    <Stack sx={{ color: "grey.500" }} spacing={2} direction="row">
+      <CircularProgress color="secondary" />
+      <CircularProgress color="success" />
+      <CircularProgress color="inherit" />
+    </Stack>
+  );
+}
+`;function f(){return r(a,{get name(){return o.name},docsName:"progress",examples:[c]})}e(f,"CircularProgressPage");export{f as default};