+++ assets/ContainerPage.16bc69e8.js
@@ -1,0 +1,15 @@
+var a=Object.defineProperty;var e=(r,i)=>a(r,"name",{value:i,configurable:!0});import{d as n,ap as m,B as s,C as o}from"./index.a47904f5.js";import{C as p}from"./ComponentInfo.6570b779.js";import"./Link.0ea0b328.js";import"./copyText.16911af4.js";import"./PaperCode.b42d056f.js";function t(){return[n(m,{}),n(o,{maxWidth:"sm",get children(){return n(s,{sx:{bgcolor:"#cfe8fc",height:"400px"}})}})]}e(t,"SimpleContainer");t.code=`import Box from "@suid/material/Box";
+import Container from "@suid/material/Container";
+import CssBaseline from "@suid/material/CssBaseline";
+
+export default function SimpleContainer() {
+  return (
+    <>
+      <CssBaseline />
+      <Container maxWidth="sm">
+        <Box sx={{ bgcolor: "#cfe8fc", height: "400px" }} />
+      </Container>
+    </>
+  );
+}
+`;function d(){return n(p,{get name(){return o.name},examples:[{bgcolor:"contrasted",component:t}]})}e(d,"ContainerPage");export{d as default};