+++ assets/ThemeProvider.8b01c8fe.js
@@ -1,0 +1,1 @@
+var n=Object.defineProperty;var r=(e,t)=>n(e,"name",{value:t,configurable:!0});import{d as o,aG as a}from"./index.a47904f5.js";function d(e){return o(a.Provider,{get value(){return e.theme},get children(){return e.children}})}r(d,"ThemeProvider");export{d as T};