+++ assets/useMediaQueryPage.cb3388c3.js
@@ -1,0 +1,8 @@
+var m=Object.defineProperty;var n=(t,e)=>m(t,"name",{value:e,configurable:!0});import{i as a,a6 as i,t as s,d as o}from"./index.a47904f5.js";import{C as p}from"./ComponentInfo.6570b779.js";import"./Link.0ea0b328.js";import"./copyText.16911af4.js";import"./PaperCode.b42d056f.js";const u=s("<span></span>");function r(){const t=a("(min-width:600px)");return(()=>{const e=u.cloneNode(!0);return i(e,()=>`(min-width:600px) matches: ${t()}`),e})()}n(r,"SimpleMediaQuery");r.code=`import useMediaQuery from "@suid/material/useMediaQuery";
+
+export default function SimpleMediaQuery() {
+  const matches = useMediaQuery("(min-width:600px)");
+
+  return <span>{\`(min-width:600px) matches: \${matches()}\`}</span>;
+}
+`;function y(){return o(p,{get name(){return a.name},docsApiName:"components/use-media-query",examples:[r],moreExamples:!1})}n(y,"useMediaQueryPage");export{y as default};