+++ assets/useMediaQueryPage.d4e31ea1.js
@@ -1,8 +1,0 @@
-var m=Object.defineProperty;var n=(t,e)=>m(t,"name",{value:e,configurable:!0});import{i as a,a6 as i,t as s,d as o}from"./index.831a2bb7.js";import{C as p}from"./ComponentInfo.e2cdfb54.js";import"./Link.8891fd7e.js";import"./copyText.1a7d19e7.js";import"./PaperCode.905e1cd4.js";const u=s("<span></span>");function r(){const t=a("(min-width:600px)");return(()=>{const e=u.cloneNode(!0);return i(e,()=>`(min-width:600px) matches: ${t()}`),e})()}n(r,"SimpleMediaQuery");r.code=`import useMediaQuery from "@suid/material/useMediaQuery";
-
-export default function SimpleMediaQuery() {
-  const matches = useMediaQuery("(min-width:600px)");
-
-  return <span>{\`(min-width:600px) matches: \${matches()}\`}</span>;
-}
-`;function y(){return o(p,{get name(){return a.name},docsApiName:"components/use-media-query",examples:[r],moreExamples:!1})}n(y,"useMediaQueryPage");export{y as default};